Successfully got rid of all delay nops from the algorithm. Now, it's all based on measuring the elapsed times between down-edges. The down-edges are detected on CA1 - and polled in the IFR register (no interrupts used for this, besides the regular T1 timer & ticks).
Longer delays (like the 18ms) are still done by polling the ticks
counter.

TBD: detect when the chip is failing to respond to the request, and
print something in that case.  After releasing the pin, if after 50us
the down edge hasn't happend, then we could abort.   That's an easy
check when spinning for that particular down edge.