The prior spinning/polling version worked fine.
But now this version captures timestamps on every CA2 downedge in the
interrupt handline, and writes them to a circular buffer, updating a
producer index in the zeropage.

So, the non-interrupt code is much simpler now.  It can actually fall
behind now and still and still catch up by walking through the pending
timestamps.

When it is caught up and needs more it calls wai and sleeps the CPU.
I guess that's the cool thing now - there is NO more spinning ever -
this is purely interrupt driven now.

The circular buffer consumes a full page of RAM.  This could be
reduced to a much smaller power of two, and even be moved inside the
zeropage but the code needed to implement the modulo-and on every
index increment would be more trouble than its worth.