Lots of updates here.
Now, a bios gets burned to ROM.  It contains a bunch of basic support,
the interrupt handler, support for serial w/rtscts, timer, lcd.
It includes wozmon & msbasic.

Programs starting anywhere at or above $400 in RAM can be loaded and
run from wozmon.   The goal is to play with programs and not have to burn
ROMs too often.