Incorporated XModem code from
http://www.6502.org/source/io/xmodem/xmodem-receive.txt

Porting was not hard at all.

But I learned about the BRK instruction and how to detect in the IRQ
handling that it was responsible for the IRQ and then how to respond.
In my case I'm happy to let it go to Wozmon straight off.

Added a 'X' hook in Wozmon to jump to XModem - when it finishes, it seems
to find its way back to Wozmon.

XModem uses the highest 2 pages of RAM for a buffer so there are no conflicts
down lower.