More updates - better understanding of the cursor move or shift command.
Future project is to have a "rendered" byte-generator which runs
faster.  SendByte currently is > 30 cycles per bit.  But if it's known
(at compile-time) that a bit is repeated, this could be reduced just
to the time to pulse the clock which is only 14 cycles.

Example is the initial address byte.  If the address is $3f, only two
changes to the dataline are needed.